Enhanced PHP installation

+++ scripts/remove_php.sh
@@ -25,6 +25,8 @@ fi
 function remove_php_fpm() {
     # PHP version.
     local PHPv="${1}"
+    local REMOVED_PHP_LOADER="${2}"
+
     if [ -z "${PHPv}" ]; then
         PHPv=${DEFAULT_PHP_VERSION:-"8.0"}
     fi
@@ -55,34 +57,12 @@ function remove_php_fpm() {
             fi
         fi
 
-        # Remove mcrypt extension.
-        #if [[ "${PHPv//.}" -lt "72" ]]; then
-        #    if "php${PHPv}" -m | grep -qw mcrypt; then
-        #        run apt-get purge -y "php${PHPv}-mcrypt"
-        #    fi
-        #elif [[ "${PHPv}" == "7.2" ]]; then
-        #    if "php${PHPv}" -m | grep -qw mcrypt; then
-                # Uninstall mcrypt pecl.
-                #run pecl uninstall mcrypt
-
-                # Unlink enabled extension.
-        #        [ -f "/etc/php/${PHPv}/cli/conf.d/20-mcrypt.ini" ] && \
-        #        run unlink "/etc/php/${PHPv}/cli/conf.d/20-mcrypt.ini"
-
-        #        [ -f "/etc/php/${PHPv}/fpm/conf.d/20-mcrypt.ini" ] && \
-        #        run unlink "/etc/php/${PHPv}/fpm/conf.d/20-mcrypt.ini"
-
-                # Remove extension.
-        #        run rm -f "/etc/php/${PHPv}/mods-available/mcrypt.ini"
-        #    fi
-        #else
-            # Use libsodium? remove separately.
-        #    info "If you're installing Libsodium extension, then remove it separately."
-        #fi
-
         # Remove PHP packages.
         # shellcheck disable=SC2046
         run apt-get purge -qq -y $(dpkg-query -l | awk '/php/ { print $2 }' | grep -wE "^php${PHPv}")
+
+        # Remove PHP loaders.
+        remove_php_loader "${PHPv}" "${REMOVED_PHP_LOADER}"
 
         # Remove PHP & FPM config files.
         warning "!! This action is not reversible !!"
@@ -125,13 +105,15 @@ function disable_ioncube_loader() {
         PHPv=${DEFAULT_PHP_VERSION:-"8.0"}
     fi
 
-    echo "Disable ionCube PHP ${PHPv} loader."
+    echo "Disable ionCube loader for PHP ${PHPv}."
 
-    [[ -f "/etc/php/${PHPv}/fpm/conf.d/05-ioncube.ini" ]] && \
-    run unlink "/etc/php/${PHPv}/fpm/conf.d/05-ioncube.ini"
+    if [[ -f "/etc/php/${PHPv}/fpm/conf.d/05-ioncube.ini" ]]; then
+        run unlink "/etc/php/${PHPv}/fpm/conf.d/05-ioncube.ini"
+    fi
 
-    [[ -f "/etc/php/${PHPv}/cli/conf.d/05-ioncube.ini" ]] && \
-    run unlink "/etc/php/${PHPv}/cli/conf.d/05-ioncube.ini"
+    if [[ -f "/etc/php/${PHPv}/cli/conf.d/05-ioncube.ini" ]]; then
+        run unlink "/etc/php/${PHPv}/cli/conf.d/05-ioncube.ini"
+    fi
 }
 
 ##
@@ -144,15 +126,15 @@ function remove_ioncube_loader() {
         PHPv=${DEFAULT_PHP_VERSION:-"8.0"}
     fi
 
-    echo "Uninstalling ionCube PHP ${PHPv} loader..."
+    echo "Uninstalling ionCube loader for PHP ${PHPv}..."
 
     disable_ioncube_loader "${PHPv}"
 
     if [[ -d /usr/lib/php/loaders/ioncube ]]; then
         run rm -fr /usr/lib/php/loaders/ioncube
-        success "ionCube PHP ${PHPv} loader has been removed."
+        success "ionCube loader for PHP ${PHPv} has been removed."
     else
-        info "ionCube PHP ${PHPv} loader couldn't be found."
+        info "ionCube loader for PHP ${PHPv} couldn't be found."
     fi
 }
 
@@ -166,13 +148,15 @@ function disable_sourceguardian_loader() {
         PHPv=${DEFAULT_PHP_VERSION:-"8.0"}
     fi
 
-    echo "Disable SourceGuardian PHP ${PHPv} loader."
+    echo "Disable SourceGuardian loader for PHP ${PHPv}."
 
-    [[ -f "/etc/php/${PHPv}/fpm/conf.d/05-sourceguardian.ini" ]] && \
-    run unlink "/etc/php/${PHPv}/fpm/conf.d/05-sourceguardian.ini"
+    if [[ -f "/etc/php/${PHPv}/fpm/conf.d/05-sourceguardian.ini" ]]; then
+        run unlink "/etc/php/${PHPv}/fpm/conf.d/05-sourceguardian.ini"
+    fi
 
-    [[ -f "/etc/php/${PHPv}/cli/conf.d/05-sourceguardian.ini" ]] && \
-    run unlink "/etc/php/${PHPv}/cli/conf.d/05-sourceguardian.ini"
+    if [[ -f "/etc/php/${PHPv}/cli/conf.d/05-sourceguardian.ini" ]]; then
+        run unlink "/etc/php/${PHPv}/cli/conf.d/05-sourceguardian.ini"
+    fi
 }
 
 ##
@@ -185,15 +169,15 @@ function remove_sourceguardian_loader() {
         PHPv=${DEFAULT_PHP_VERSION:-"8.0"}
     fi
 
-    echo "Uninstalling SourceGuardian PHP ${PHPv} loader..."
+    echo "Uninstalling SourceGuardian loader for PHP ${PHPv}..."
 
     disable_sourceguardian_loader "${PHPv}"
 
     if [[ -d /usr/lib/php/loaders/sourceguardian ]]; then
         run rm -fr /usr/lib/php/loaders/sourceguardian
-        success "SourceGuardian PHP ${PHPv} loader has been removed."
+        success "SourceGuardian loader for PHP ${PHPv} has been removed."
     else
-        info "SourceGuardian PHP ${PHPv} loader couldn't be found."
+        info "SourceGuardian loader for PHP ${PHPv} couldn't be found."
     fi
 }
 
@@ -278,18 +262,25 @@ function init_php_fpm_removal() {
 
     eval set -- "${OPTS}"
 
-    while true
-    do
+    while true; do
         case "${1}" in
-            -p | --php-version) shift
-                OPT_PHP_VERSIONS+=("${1}")
+            -p | --php-version) 
+                shift
+                if [[ "${1}" == "all" ]]; then
+                    # Include versions from config file.
+                    read -r -a OPT_PHP_VERSIONS <<< "${PHP_VERSIONS}"
+                else
+                    OPT_PHP_VERSIONS+=("${1}")
+                fi
                 shift
             ;;
-            -l | --php-loader) shift
+            -l | --php-loader) 
+                shift
                 OPT_PHP_LOADER="${1}"
                 shift
             ;;
-            --) shift
+            --) 
+                shift
                 break
             ;;
             *)
@@ -298,9 +289,6 @@ function init_php_fpm_removal() {
             ;;
         esac
     done
-
-    # Include versions from config file.
-    read -r -a REMOVED_PHP_VERSIONS <<< "${PHP_VERSIONS}"
 
     if [[ "${#OPT_PHP_VERSIONS[@]}" -gt 0 ]]; then
         REMOVED_PHP_VERSIONS+=("${OPT_PHP_VERSIONS[@]}")
@@ -388,9 +376,8 @@ function init_php_fpm_removal() {
 
     # Remove all selected PHP versions.
     if [[ "${#REMOVED_PHP_VERSIONS[@]}" -gt 0 ]]; then
-        for PHPV in "${REMOVED_PHP_VERSIONS[@]}"; do
-            remove_php_fpm "${PHPV}"
-            remove_php_loader "${PHPV}" "${OPT_PHP_LOADER}"
+        for PHP_VER in "${REMOVED_PHP_VERSIONS[@]}"; do
+            remove_php_fpm "${PHP_VER}" "${OPT_PHP_LOADER}"
             echo ""
         done
 
@@ -398,13 +385,18 @@ function init_php_fpm_removal() {
         if [[ "${DRYRUN}" != true ]]; then
             # New logic for multiple PHP removal in batch.
             PHP_IS_EXISTS=false
-            for PHPV in "${REMOVED_PHP_VERSIONS[@]}"; do
-                [[ -n $(command -v "${PHPV}") ]] && PHP_IS_EXISTS=true
+            for PHP_VER in "${REMOVED_PHP_VERSIONS[@]}"; do
+                [[ -n $(command -v "${PHP_VER}") ]] && PHP_IS_EXISTS=true
             done
 
             if [[ "${PHP_IS_EXISTS}" == false ]]; then
                 echo "Removing additional unused PHP packages..."
                 run apt-get purge -qq -y dh-php php-common php-pear php-xml pkg-php-tools fcgiwrap spawn-fcgi
+
+                # Remove openswoole official repository.
+                if echo "${PHP_EXTENSIONS}" | grep -qwE "openswoole"; then
+                    run add-apt-repository -y ppa:openswoole/ppa
+                fi
 
                 # Remove PHP repository.
                 run add-apt-repository -y --remove ppa:ondrej/php
@@ -446,8 +438,8 @@ if [[ -n $(command -v php5.6) || \
     if [[ "${REMOVE_PHP}" == Y* || "${REMOVE_PHP}" == y* ]]; then
         init_php_fpm_removal "$@"
     else
-        echo "Found PHP package, but not removed."
+        echo "Found PHP packages, but not removed."
     fi
 else
-    info "Oops, PHP package installation not found."
+    info "Oops, PHP packages installation not found."
 fi
